requirements.txt
flask
Flask-SQLAlchemy
requests
python-dotenv